+++ console/src/modules/system/settings/module.ts
@@ -1,31 +1,30 @@
 import { definePlugin } from "@halo-dev/console-shared";
-import SystemSettingsLayout from "./layouts/SystemSettingsLayout.vue";
-import SystemSetting from "./SystemSetting.vue";
+import SystemSettings from "./SystemSettings.vue";
 import { IconSettings } from "@halo-dev/components";
 import { markRaw } from "vue";
+import BasicLayout from "@/layouts/BasicLayout.vue";
 
 export default definePlugin({
   components: {},
   routes: [
     {
       path: "/settings",
-      component: SystemSettingsLayout,
-      redirect: "/settings/basic",
-      meta: {
-        title: "core.setting.title",
-        permissions: ["system:settings:view"],
-        menu: {
-          name: "core.sidebar.menu.items.settings",
-          group: "system",
-          icon: markRaw(IconSettings),
-          priority: 2,
-        },
-      },
+      component: BasicLayout,
       children: [
         {
-          path: ":group",
+          path: "",
           name: "SystemSetting",
-          component: SystemSetting,
+          component: SystemSettings,
+          meta: {
+            title: "core.setting.title",
+            permissions: ["system:settings:view"],
+            menu: {
+              name: "core.sidebar.menu.items.settings",
+              group: "system",
+              icon: markRaw(IconSettings),
+              priority: 2,
+            },
+          },
         },
       ],
     },
